+++ src/main.py
@@ -5,7 +5,6 @@
 # Date of release: 13.06.2024
 # Last Edited: 20.06.2024
 
-
 """
 This script facilitates user login and session initialization for PokeRogue. It offers a menu-driven interface to 
 perform various account and game data actions after a successful login using either requests or Selenium.
@@ -31,32 +30,14 @@ Modules:
 """
 
 import getpass
-# Securely obtains the password from the user without echoing.
-
 import requests
-# Handles HTTP sessions and requests for logging in and maintaining session state.
-
 import brotli  # noqa: F401
-# Provides Brotli compression. Not directly used in this script but might be a dependency for other modules.
 
-from modules import requestsLogic, Rogue,  SeleniumLogic, config
-# requestLogic: Handles the login logic using requests.
-# Rogue: Initializes and interacts with the PokeRogue session.
-# SeleniumLogic: Handles login logic using Selenium for browser-based interactions.
-# Config: Provides configuration settings and update checking functionality.
-
+from modules import requestsLogic, Rogue, SeleniumLogic, config
+from modules.handler import OperationSuccessful, handle_operation_exceptions
 from colorama import Fore, Style, init
-# Fore, Style: Used for terminal text color formatting.
-# init: Initializes colorama for colored console output.
-
 from utilities import cFormatter, Color, CustomLogger
-# cFormatter: Custom formatter for colored printing and logging.
-# Color: Enumeration defining color codes for cFormatter.
-# Custom Logger: Provides custom logging functionality for the script.
-
-
 from datetime import datetime, timedelta
-# datetime, timedelta: For date and time operations, particularly for update checking.
 
 init()
 config.f_initFolders()
@@ -64,9 +45,99 @@ logger = CustomLogger()
 config.f_checkForUpdates(requests, datetime, timedelta, Style)
 config.f_printWelcomeText()
 
+# Global list for pre-command messages
+pre_command_messages = []
+
+def m_propagateMessages():
+    """
+    Prints messages from the global pre_command_messages list and then clears the list.
+    """
+    global pre_command_messages
+    if pre_command_messages:
+        for message in pre_command_messages:
+            cFormatter.print(Color.INFO, message)
+        pre_command_messages.clear()
+
+@handle_operation_exceptions
+def m_executeOptions(choice_index, valid_choices):
+    m_propagateMessages()  # Print messages before executing the option
+    for idx, func in valid_choices:
+        if idx == choice_index:
+            func()
+            break
+        elif idx == 'exit':
+            raise KeyboardInterrupt
+
+@handle_operation_exceptions
+def m_mainMenu(rogue):
+    useWhenDone = f'{Fore.LIGHTYELLOW_EX}(Use when Done)'
+    title = f'{config.title}>'
+
+    term = [
+        (title, 'title'),
+        ('Account Actions', 'category'),
+        (('Create a backup', ''), rogue.create_backup),
+        (('Recover your backup', ''), rogue.restore_backup),
+        (('Load Game-Data from server', ''), rogue.get_trainer_data),
+        (('Change save-slot to edit', 'workingWIP'), rogue.f_changeSaveSlot),
+        (('Edit account stats', 'workingWIP'), rogue.f_editAccountStats),
+
+        ('Trainer Data Actions', 'category'),
+        (('Edit a starter', ''), rogue.edit_starter_separate),
+        (('Edit your egg-tickets', ''), rogue.add_ticket),
+        (('Edit candies on a starter', ''), rogue.f_addCandies),
+        (('Edit Egg-hatch durations', 'workingWIP'), rogue.f_editHatchWaves),
+        (('Generate eggs', 'workingWIP'), rogue.f_addEggsGenerator),
+        (('Unlock all vouchers', ''), rogue.f_editVouchers),
+        (('Unlock all starters', ''), rogue.f_unlockStarters),
+        (('Unlock all achievements', ''), rogue.f_editAchivements),
+        (('Unlock all gamemodes', ''), rogue.f_editGamemodes),
+        (('Unlock Everything', ''), rogue.f_unlockAllCombined),
+
+        ('Session Data Actions', 'category'),
+        (('Edit CURRENT Pokemon Party', ''), rogue.edit_pokemon_party),
+        (('Edit money amount', 'workingWIP'), rogue.f_editMoney),
+        (('Edit pokeballs amount', 'workingWIP'), rogue.f_editPokeballs),
+        (('Edit current biome', 'workingWIP'), rogue.f_editBiome),
+        (('Edit Items', f'{Fore.GREEN + Style.BRIGHT}NEW{Style.RESET_ALL}'), rogue.f_submenuItemEditor),
+
+        ('Print game information', 'category'),
+        (('Show all Pokemon ID', 'workingWIP'), rogue.print_pokedex),
+        (('Show all Biome IDs', 'workingWIP'), rogue.f_printBiomes),
+        (('Show all Move IDs', 'workingWIP'), rogue.print_moves),
+        (('Show all Vouchers IDs', 'workingWIP'), rogue.print_vouchers),
+        (('Show all Natures IDs', 'workingWIP'), rogue.print_natures),
+        (('Show all NaturesSlot IDs', 'workingWIP'), rogue.print_natureSlot),
+
+        ('You can always edit your JSON manually as well!', 'helper'),
+        (('Save data and upload to the Server', useWhenDone), rogue.update_all),
+        (('Print help and program information', ''), config.f_printHelp),
+        (('Logout', ''), rogue.logout),
+        (title, 'title'),
+    ]
+
+    try:
+        while True:
+            validChoices = cFormatter.m_initializeMenu(term)
+            userInput = input('Command: ').strip().lower()
+
+            if userInput == 'exit':
+                raise KeyboardInterrupt
+            
+            if userInput.isdigit():
+                choiceIndex = int(userInput)
+                m_executeOptions(choiceIndex, validChoices)
+            else:
+                cFormatter.print(Color.INFO, 'Invalid input. Please enter a number.')
+    except OperationSuccessful as os:
+        cFormatter.print(Color.DEBUG, f'Operation successful: {os}')
+    except KeyboardInterrupt:
+        cFormatter.print(Color.DEBUG, '\nProgram interrupted by user.')
+        exit()
+
+@handle_operation_exceptions
 def main():
     while True:
-        # Try loginChoice
         try:
             loginChoice = input('Please choose a method of logging in: ')
             loginChoice = int(loginChoice)  # Attempt to convert input to integer
@@ -81,38 +152,35 @@ def main():
             cFormatter.print(Color.DEBUG, 'Invalid input. Please enter a number.')
             continue  # Prompt user again for input
 
-        # Try login
         try:
             username = input('Username: ')
             password = getpass.getpass('Password (password is hidden): ')
 
             session = requests.Session()
-            # When using requests
             if loginChoice == 1:
                 login = requestsLogic(username, password)
                 try:
                     if login.login():
                         cFormatter.print(Color.INFO, f'Logged in as: {config.f_anonymizeName(username)}')
-                        session.cookies.set('pokerogue_sessionId', login.session_id, domain='pokerogue.net')
-                        rogue = Rogue(session, login.token, login.session_id)
+                        session.cookies.set('pokerogue_sessionId', login.sessionId, domain='pokerogue.net')
+                        rogue = Rogue(session, login.token, login.sessionId)
                         break
                 except Exception as e:
                     cFormatter.print(Color.CRITICAL, f'Something went wrong. {e}', isLogging=True)
 
-            # When using browser
             elif loginChoice in [2, 3]:
                 try:
                     if loginChoice == 3:
                         cFormatter.print(Color.INFO, 'Do not close your browser and do not browse in the game!')
                         cFormatter.print(Color.INFO, 'Do not close your browser and do not browse in the game!')
                         cFormatter.print(Color.INFO, 'Do not close your browser and do not browse in the game!')
-                    selenium_logic = SeleniumLogic(username, password, 120, useScripts=(loginChoice == 3))
-                    session_id, token, driver = selenium_logic.logic()
+                    seleniumLogic = SeleniumLogic(username, password, 120, useScripts=(loginChoice == 3))
+                    sessionId, token, driver = seleniumLogic.logic()
 
-                    if session_id and token and driver:
-                        cFormatter.print(Color.INFO, f'Logged in as: {username.capitalize()}')
-                        session.cookies.set('pokerogue_sessionId', session_id, domain='pokerogue.net')
-                        rogue = Rogue(session, auth_token=token, clientSessionId=session_id, driver=driver, useScripts=(loginChoice == 3))
+                    if sessionId and token and driver:
+                        cFormatter.print(Color.INFO, f'Logged in as: {config.f_anonymizeName(username)}')
+                        session.cookies.set('pokerogue_sessionId', sessionId, domain='pokerogue.net')
+                        rogue = Rogue(session, auth_token=token, clientSessionId=sessionId, driver=driver, useScripts=(loginChoice == 3))
                         break
                     else:
                         cFormatter.print(Color.CRITICAL, 'Failed to retrieve necessary authentication data from Selenium.')
@@ -125,78 +193,14 @@ def main():
             cFormatter.print(Color.DEBUG, '\nProgram interrupted by user.')
             exit()
 
-    # Define menu variables and our menu
-    useWhenDone = f'{Fore.LIGHTYELLOW_EX}(Use when Done)'
-    title = f'{config.title}>'
+    m_mainMenu(rogue)
 
-    # See cFormatter.initialize_menu() for more information
-    term = [
-        (title, 'title'),
-        ('Account Actions', 'category'),
-        (('Create a backup', ''), rogue.create_backup),
-        (('Recover your backup', ''), rogue.restore_backup),
-        (('Load Game-Data from server', ''), rogue.get_trainer_data),
-        (('Change save-slot to edit', ''), rogue.f_changeSaveSlot),
-        (('Edit account stats', ''), rogue.f_editAccountStats),
-
-        ('Trainer Data Actions', 'category'),
-        (('Edit a starter', ''), rogue.edit_starter_separate),
-        (('Edit your egg-tickets', ''), rogue.add_ticket),
-        (('Edit candies on a starter', ''), rogue.f_addCandies),
-        (('Edit Egg-hatch durations', ''), rogue.f_editHatchWaves),
-        (('Generate eggs', ''), rogue.f_addEggsGenerator),
-        (('Unlock all vouchers', ''), rogue.f_editVouchers),
-        (('Unlock all starters', ''), rogue.f_unlockStarters),
-        (('Unlock all achievements', ''), rogue.f_editAchivements),
-        (('Unlock all gamemodes', ''), rogue.f_editGamemodes),
-        (('Unlock Everything', ''), rogue.f_unlockAllCombined),
-
-        ('Session Data Actions', 'category'),
-        (('Edit CURRENT Pokemon Party', ''), rogue.edit_pokemon_party),
-        (('Edit money amount', ''), rogue.f_editMoney),
-        (('Edit pokeballs amount', ''), rogue.f_editPokeballs),
-        (('Edit current biome', ''), rogue.f_editBiome),
-        (('Edit Items', f'{Fore.GREEN + Style.BRIGHT}NEW{Style.RESET_ALL}'), rogue.f_submenuItemEditor),
-
-        ('Print game information', 'category'),
-        (('Show all Pokemon ID', ''), rogue.print_pokedex),
-        (('Show all Biome IDs', ''), rogue.print_biomes),
-        (('Show all Move IDs', ''), rogue.print_moves),
-        (('Show all Vouchers IDs', ''), rogue.print_vouchers),
-        (('Show all Natures IDs', ''), rogue.print_natures),
-        (('Show all NaturesSlot IDs', ''), rogue.print_natureSlot),
-
-        ('You can always edit your JSON manually as well!', 'helper'),
-        (('Save data and upload to the Server', useWhenDone), rogue.update_all),
-        (('Print help and program information', ''), config.f_printHelp),
-        (('Logout', ''), rogue.logout),
-        (title, 'title'),
-    ]
-
-    try:
-
-        while True:
-            valid_choices = cFormatter.initialize_menu(term)
-            user_input = input('Command: ').strip().lower()
-
-            if user_input == 'exit':
-                raise KeyboardInterrupt
-            
-            if user_input.isdigit():
-                choice_index = int(user_input)
-                for idx, func in valid_choices:
-                    if idx == choice_index:
-                        func()
-                        break
-                    elif idx == 'exit':
-                        raise KeyboardInterrupt
-                else:
-                    cFormatter.print(Color.INFO, 'Invalid selection. Please choose a valid menu option.')
-            else:
-                cFormatter.print(Color.INFO, 'Invalid input. Please enter a number.')
-
-    except KeyboardInterrupt:
-        cFormatter.print(Color.DEBUG, '\nProgram interrupted by user.')
-        exit()
 if __name__ == '__main__':
-    main()
+    while True:
+        try:
+            main()
+        except OperationSuccessful as os:
+            cFormatter.print(Color.DEBUG, f'Operation successful: {os}')
+        except KeyboardInterrupt:
+            cFormatter.print(Color.DEBUG, '\nProgram interrupted by user.')
+            exit()